feat(auth): add loading and pre-login components for improved user experience during authentication
